some refactoring to reuse code to escape/unescape double quotes
